fix policy evaluation test and add useful debug tools for the future
the test assumed that the transitions follow a certain structure and hardcoded this structure into the test results and setup(!). pymimir v2 does behave differently on different OSs though and the order that was assumed does not always hold. It has been made a little more robust to such changes now.